checkpoint

+++ src/preprocessing/averaging.py
@@ -1,5 +1,5 @@
 import numpy
-from src.util.metrics import capture_timing_info
+from util.metrics import capture_timing_info
 
 
 @capture_timing_info()
@@ -15,4 +15,4 @@ def calculate_profile(images: list[numpy.ndarray], smooth_steps: int = 0) -> num
 
 @capture_timing_info()
 def apply_profile_to_image(image: numpy.ndarray, profile: numpy.ndarray) -> numpy.ndarray:
-    return numpy.clip(image / profile, 0, 255)
+    return numpy.clip(image / profile, 0, 255).astype(numpy.uint8)
